+++ crates/matrix-sdk/src/sliding_sync/mod.rs
@@ -342,6 +342,11 @@ impl SlidingSync {
 
             // Update the lists.
             let updated_lists = {
+                debug!(
+                    lists = ?sliding_sync_response.lists,
+                    "Update lists"
+                );
+
                 let mut updated_lists = Vec::with_capacity(sliding_sync_response.lists.len());
                 let mut lists = self.inner.lists.write().await;
 
@@ -694,6 +699,16 @@ impl SlidingSync {
     /// Force caching the current sliding sync to storage.
     pub async fn force_cache_to_storage(&self, to_device_token: Option<String>) -> Result<()> {
         self.cache_to_storage(to_device_token).await
+    }
+
+    /// Read the static extension configuration for this Sliding Sync.
+    ///
+    /// Note: this is not the next content of the sticky parameters, but rightly
+    /// the static configuration that was set during creation of this
+    /// Sliding Sync.
+    pub fn extensions_config(&self) -> ExtensionsConfig {
+        let sticky = self.inner.sticky.read().unwrap();
+        sticky.data().extensions.clone()
     }
 }